Centralize x-* OpenAPI extension parsing in OpenApiExtensionReader
x-tagGroups, x-displayName, x-namespace/x-api-name, x-beta and
x-codeSamples were each parsed inline where needed (OpenApiGenerator,
OperationViewModel, OperationPageModel) with the JsonNodeExtension
unwrapping repeated every time. One reader now owns all of it.
